+++ vector/src/main/java/im/vector/riotx/features/home/room/detail/composer/ComposerEditText.kt
@@ -1,0 +1,58 @@
+/*
+ * Copyright 2019 New Vector Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package im.vector.riotx.features.home.room.detail.composer
+
+import android.content.Context
+import android.net.Uri
+import android.os.Build
+import android.util.AttributeSet
+import android.view.inputmethod.EditorInfo
+import android.view.inputmethod.InputConnection
+import androidx.appcompat.widget.AppCompatEditText
+import androidx.core.view.inputmethod.EditorInfoCompat
+import androidx.core.view.inputmethod.InputConnectionCompat
+
+class ComposerEditText @JvmOverloads constructor(context: Context, attrs: AttributeSet? = null, defStyleAttr: Int = android.R.attr.editTextStyle)
+    : AppCompatEditText(context, attrs, defStyleAttr) {
+
+    interface Callback {
+        fun onRichContentSelected(contentUri: Uri): Boolean
+    }
+
+    var callback: Callback? = null
+
+    override fun onCreateInputConnection(editorInfo: EditorInfo): InputConnection {
+        val ic: InputConnection = super.onCreateInputConnection(editorInfo)
+        EditorInfoCompat.setContentMimeTypes(editorInfo, arrayOf("*/*"))
+
+        val callback =
+                InputConnectionCompat.OnCommitContentListener { inputContentInfo, flags, _ ->
+                    val lacksPermission = (flags and
+                            InputConnectionCompat.INPUT_CONTENT_GRANT_READ_URI_PERMISSION) != 0
+                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N_MR1 && lacksPermission) {
+                        try {
+                            inputContentInfo.requestPermission()
+                        } catch (e: Exception) {
+                            return@OnCommitContentListener false
+                        }
+                    }
+                    callback?.onRichContentSelected(inputContentInfo.contentUri) ?: false
+                }
+        return InputConnectionCompat.createWrapper(ic, editorInfo, callback)
+    }
+}