+++ vector/src/main/java/im/vector/riotx/features/home/room/list/RoomListFragment.kt
@@ -41,13 +41,15 @@ import im.vector.riotx.core.platform.StateView
 import im.vector.riotx.core.platform.VectorBaseFragment
 import im.vector.riotx.features.home.room.list.widget.FabMenuView
 import im.vector.riotx.features.notifications.NotificationDrawerManager
+import im.vector.riotx.features.share.SharedData
 import kotlinx.android.parcel.Parcelize
 import kotlinx.android.synthetic.main.fragment_room_list.*
 import javax.inject.Inject
 
 @Parcelize
 data class RoomListParams(
-        val displayMode: RoomListFragment.DisplayMode
+        val displayMode: RoomListFragment.DisplayMode,
+        val sharedData: SharedData? = null
 ) : Parcelable
 
 class RoomListFragment : VectorBaseFragment(), RoomSummaryController.Listener, OnBackPressed, FabMenuView.Listener {
@@ -56,7 +58,8 @@ class RoomListFragment : VectorBaseFragment(), RoomSummaryController.Listener, O
         HOME(R.string.bottom_action_home),
         PEOPLE(R.string.bottom_action_people_x),
         ROOMS(R.string.bottom_action_rooms),
-        FILTERED(/* Not used */ R.string.bottom_action_rooms)
+        FILTERED(/* Not used */ 0),
+        SHARE(/* Not used */ 0)
     }
 
     companion object {
@@ -106,7 +109,12 @@ class RoomListFragment : VectorBaseFragment(), RoomSummaryController.Listener, O
         setupRecyclerView()
         roomListViewModel.subscribe { renderState(it) }
         roomListViewModel.openRoomLiveData.observeEventFirstThrottle(this, 800L) {
-            navigator.openRoom(requireActivity(), it)
+            if (roomListParams.displayMode == DisplayMode.SHARE) {
+                val sharedData = roomListParams.sharedData ?: return@observeEventFirstThrottle
+                navigator.openRoomForSharing(requireActivity(), it, sharedData)
+            } else {
+                navigator.openRoom(requireActivity(), it)
+            }
         }
 
         createChatFabMenu.listener = this
@@ -121,10 +129,10 @@ class RoomListFragment : VectorBaseFragment(), RoomSummaryController.Listener, O
 
     private fun setupCreateRoomButton() {
         when (roomListParams.displayMode) {
-            DisplayMode.HOME     -> createChatFabMenu.isVisible = true
-            DisplayMode.PEOPLE   -> createChatRoomButton.isVisible = true
-            DisplayMode.ROOMS    -> createGroupRoomButton.isVisible = true
-            DisplayMode.FILTERED -> Unit // No button in this mode
+            DisplayMode.HOME   -> createChatFabMenu.isVisible = true
+            DisplayMode.PEOPLE -> createChatRoomButton.isVisible = true
+            DisplayMode.ROOMS  -> createGroupRoomButton.isVisible = true
+            else               -> Unit // No button in this mode
         }
 
         createChatRoomButton.setOnClickListener {
@@ -147,10 +155,10 @@ class RoomListFragment : VectorBaseFragment(), RoomSummaryController.Listener, O
                             RecyclerView.SCROLL_STATE_DRAGGING,
                             RecyclerView.SCROLL_STATE_SETTLING -> {
                                 when (roomListParams.displayMode) {
-                                    DisplayMode.HOME     -> createChatFabMenu.hide()
-                                    DisplayMode.PEOPLE   -> createChatRoomButton.hide()
-                                    DisplayMode.ROOMS    -> createGroupRoomButton.hide()
-                                    DisplayMode.FILTERED -> Unit
+                                    DisplayMode.HOME   -> createChatFabMenu.hide()
+                                    DisplayMode.PEOPLE -> createChatRoomButton.hide()
+                                    DisplayMode.ROOMS  -> createGroupRoomButton.hide()
+                                    else               -> Unit
                                 }
                             }
                         }
@@ -187,10 +195,10 @@ class RoomListFragment : VectorBaseFragment(), RoomSummaryController.Listener, O
     private val showFabRunnable = Runnable {
         if (isAdded) {
             when (roomListParams.displayMode) {
-                DisplayMode.HOME     -> createChatFabMenu.show()
-                DisplayMode.PEOPLE   -> createChatRoomButton.show()
-                DisplayMode.ROOMS    -> createGroupRoomButton.show()
-                DisplayMode.FILTERED -> Unit
+                DisplayMode.HOME   -> createChatFabMenu.show()
+                DisplayMode.PEOPLE -> createChatRoomButton.show()
+                DisplayMode.ROOMS  -> createGroupRoomButton.show()
+                else               -> Unit
             }
         }
     }
@@ -235,7 +243,7 @@ class RoomListFragment : VectorBaseFragment(), RoomSummaryController.Listener, O
                 }
                 .isNullOrEmpty()
         val emptyState = when (roomListParams.displayMode) {
-            DisplayMode.HOME     -> {
+            DisplayMode.HOME   -> {
                 if (hasNoRoom) {
                     StateView.State.Empty(
                             getString(R.string.room_list_catchup_welcome_title),
@@ -249,19 +257,19 @@ class RoomListFragment : VectorBaseFragment(), RoomSummaryController.Listener, O
                             getString(R.string.room_list_catchup_empty_body))
                 }
             }
-            DisplayMode.PEOPLE   ->
+            DisplayMode.PEOPLE ->
                 StateView.State.Empty(
                         getString(R.string.room_list_people_empty_title),
                         ContextCompat.getDrawable(requireContext(), R.drawable.ic_home_bottom_chat),
                         getString(R.string.room_list_people_empty_body)
                 )
-            DisplayMode.ROOMS    ->
+            DisplayMode.ROOMS  ->
                 StateView.State.Empty(
                         getString(R.string.room_list_rooms_empty_title),
                         ContextCompat.getDrawable(requireContext(), R.drawable.ic_home_bottom_group),
                         getString(R.string.room_list_rooms_empty_body)
                 )
-            DisplayMode.FILTERED ->
+            else               ->
                 // Always display the content in this mode, because if the footer
                 StateView.State.Content
         }